requirements.txt
# Core Frameworks
fastapi
uvicorn
streamlit
python-dotenv

# LangChain and Google AI
langchain
langchain-core
langchain-google-genai
langgraph

# Data Handling
pandas